shift the getInfoByProjectUrl function in commonFunctions.js file
